renamed 'ftrack_module.py' > 'ftrack_addon.py'
